few chnages and cleanup

+++ notes/installer.py
@@ -5,36 +5,51 @@ import os
 import argparse
 
 
-def copy_item(source, destination):
+def copy_item(source, destination, permissions, mode):
     if os.path.isfile(source):
         # If the source is a file, use copy_file
-        copy_file(source, destination)
+        copy_file(source, destination, permissions, mode)
         return "File"
     elif os.path.isdir(source):
         # If the source is a directory, use copy_tree
-        copy_tree(source, destination)
-        return "Folder"
+        #copy_tree(source, destination,  permissions, mode)
+        return "Folders are not supported"
     
     return "Invalid source path"
     
 
-def copy_file(source, destination):
-    try:
-        shutil.copy2(source, destination)
-    except IOError as e:
-        if e.errno != errno.ENOENT:
-            raise
-        # try creating parent directories
-        os.makedirs(os.path.dirname(destination))
-        shutil.copy2(source, destination)
+def copy_file(source, destination, permissions, mode):
 
-def copy_tree(source, destination):
-    try:
-        shutil.rmtree(destination)  # Remove the destination folder if it exists
-    except OSError:
-        pass  # Ignore errors if the destination folder doesn't exist
+    if os.path.exists(destination) and os.path.isfile(destination) and mode=="install":
+        dest_permissions = os.stat(destination).st_mode
+        with open(source, 'rb') as source_file:
+            content = source_file.read()
+        with open(destination, 'wb') as dest_file:
+            dest_file.write(content)
+        os.chmod(destination, dest_permissions)
 
-    shutil.copytree(source, destination)
+    else:
+        try:
+            shutil.copy2(source, destination)
+        except IOError as e:
+            if e.errno != errno.ENOENT:
+                raise
+            # try creating parent directories
+            os.makedirs(os.path.dirname(destination))
+            shutil.copy2(source, destination)
+    
+    if permissions is not None and mode=="install":
+        os.chmod(destination, int(permissions, base=8))
+    
+
+
+# def copy_tree(source, destination,  permissions, mode):
+#     try:
+#         shutil.rmtree(destination)  # Remove the destination folder if it exists
+#     except OSError:
+#         pass  # Ignore errors if the destination folder doesn't exist
+
+#     shutil.copytree(source, destination)
 
 def print_pretty_table(data):
     col_widths = [max(len(str(item)) for item in col) for col in zip(*data)]
@@ -65,17 +80,19 @@ def main():
 
 
     if args.backup:
-        # Execute Backup of all files:
         backup_folder = instructions['backup_folder']
         data = []
-        
+        mode = "backup"
         print("# Backup")
 
         for manipulation in instructions['release']:
             source = manipulation['destination']
+            permissions = manipulation.get('permissions', None)
+            status=""
+            type=""
 
             try:
-                type = copy_item(source,backup_folder +"/"+source)
+                type = copy_item(source,backup_folder +"/"+source, permissions, mode)
                 status = "Success" if type != "Invalid source path" else "Skipped"
             except Exception as e:
                 status = repr(e)
@@ -86,17 +103,22 @@ def main():
 
     if args.install:
         data = []
+        mode = "install"
         print("# Installation")
 
         for manipulation in instructions['release']:
             source = manipulation['source']
             destination =  manipulation['destination']
+            permissions = manipulation.get('permissions', None)
+            status=""
+            type=""
 
             try:
-                type = copy_item(source,destination)
+                type = copy_item(source,destination, permissions, mode)
                 status = "Success"
             except Exception as e:
                 status = repr(e)
+
             new_row = ["Installation", source, destination, type, status]
             data.append(new_row)
         print_pretty_table(data)
